:art: refactor : 수익률 계산 로직을 다른 클래스로 추출
