+++ src/test/java/lotto/step2/domain/LottoResultCalculatorTest.java
@@ -1,0 +1,17 @@
+package lotto.step2.domain;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+import java.util.List;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+public class LottoResultCalculatorTest {
+    @Test
+    @DisplayName("수익률 계산하기")
+    void calculate_yield() {
+        double yield = LottoResultCalculator.parseYield(List.of(LottoRank.FIFTH), PaymentPriceTest.PAYMENT_PRICE);
+        assertThat(yield).isEqualTo(0.35);
+    }
+}
